refactor: separate api-version types, constants, and classes into appropriate directories
Move constants from types to utils and APIVersionManager class to client directory
to maintain better architectural separation of concerns. Types directory now contains
only type definitions (enums, interfaces, type aliases).
