+++ .gitignore
@@ -26,6 +26,7 @@ coverage
 *.njsproj
 *.sln
 *.sw?
+.serena
 
 # Environment variables
 .env
