+++ docs/enumerations/APIVersion.md
@@ -6,12 +6,9 @@
 
 # Enumeration: APIVersion
 
-Defined in: [types/api-version.ts:8](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L8)
+Defined in: [types/api-version.ts:11](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L11)
 
-API Version Management Types for X Ads SDK
-
-X Ads API follows a 6-month deprecation cycle with 2 concurrent versions supported.
-This module provides type-safe version management.
+Supported API versions for X (Twitter) Ads API
 
 ## Enumeration Members
 
@@ -19,7 +16,7 @@ This module provides type-safe version management.
 
 > **V11**: `"11.0"`
 
-Defined in: [types/api-version.ts:9](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L9)
+Defined in: [types/api-version.ts:12](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L12)
 
 ***
 
@@ -27,4 +24,4 @@ Defined in: [types/api-version.ts:9](https://github.com/kage1020/x-ads-sdk/blob/
 
 > **V12**: `"12.0"`
 
-Defined in: [types/api-version.ts:10](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L10)
+Defined in: [types/api-version.ts:13](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L13)
