+++ docs/type-aliases/SupportedAPIVersion.md
@@ -6,6 +6,8 @@
 
 # Type Alias: SupportedAPIVersion
 
-> **SupportedAPIVersion** = *typeof* [`SUPPORTED_VERSIONS`](../variables/SUPPORTED_VERSIONS.md)\[`number`\]
+> **SupportedAPIVersion** = [`V11`](../enumerations/APIVersion.md#v11) \| [`V12`](../enumerations/APIVersion.md#v12)
 
-Defined in: [types/api-version.ts:18](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L18)
+Defined in: [types/api-version.ts:20](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L20)
+
+Type for supported API versions
