+++ docs/variables/API_VERSION_METADATA.md
@@ -8,4 +8,6 @@
 
 > `const` **API\_VERSION\_METADATA**: `Record`\<[`APIVersion`](../enumerations/APIVersion.md), [`APIVersionInfo`](../interfaces/APIVersionInfo.md)\>
 
-Defined in: [types/api-version.ts:29](https://github.com/kage1020/x-ads-sdk/blob/main/src/types/api-version.ts#L29)
+Defined in: utils/api-version.ts:6
+
+Metadata for all API versions
