+++ package.json
@@ -81,7 +81,6 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.2.2",
-    "@rollup/plugin-node-resolve": "^16.0.1",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^12.1.4",
     "@types/node": "^24.3.0",
