+++ src/client/index.ts
@@ -1,3 +1,4 @@
+export * from './api-version-manager.js';
 export * from './base.js';
 export * from './rate-limit.js';
 export * from './retry.js';
