+++ src/index.ts
@@ -4,6 +4,5 @@ export * from './errors/index.js';
 export * from './modules/index.js';
 export * from './paginators/index.js';
 export * from './plugins/index.js';
-export * from './types/api-version.js';
 export * from './types/index.js';
 export * from './utils/index.js';
